backend/requirements.txt
psycopg2-binary==2.9.9
pytest==8.1.1
yfinance>=0.2.37
pandas>=2.2.1
matplotlib>=3.8.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0